削除前の確認ダイアログを作成

+++ flutter/lib/components/cards/event_card.dart
@@ -45,12 +45,33 @@ class EventCard extends HookWidget {
                   },
                 ),
                 IconButton(
-                  icon: Icon(Icons.delete),
-                  onPressed: () async {
-                    await database.deleteEvent(event);
-                    _eventListPv.refresh();
-                  },
-                ),
+                    icon: Icon(Icons.delete),
+                    onPressed: () {
+                      showDialog<int>(
+                          context: context,
+                          barrierDismissible: false,
+                          builder: (BuildContext context) {
+                            return AlertDialog(
+                              title: Text('イベントの削除'),
+                              content: Text(
+                                  '${dateFormat(event.date)} の ${event.name} イベントを削除してよろしいですか？'),
+                              actions: <Widget>[
+                                ElevatedButton(
+                                  child: Text('Cancel'),
+                                  onPressed: () => Navigator.of(context).pop(0),
+                                ),
+                                ElevatedButton(
+                                  child: Text('OK'),
+                                  onPressed: () async {
+                                    await database.deleteEvent(event);
+                                    _eventListPv.refresh();
+                                    Navigator.of(context).pop(0);
+                                  },
+                                ),
+                              ],
+                            );
+                          });
+                    }),
                 const SizedBox(width: 8),
               ],
             ),
